Se añade lista y detalle de categorias

+++ app/src/main/java/com/arribas/myshoppinglist/ui/view/screen/shoplist/shoplistDetail/ShoplistDetailViewModel.kt
@@ -7,18 +7,15 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.arribas.myshoppinglist.data.model.ArticleShop
 import com.arribas.myshoppinglist.data.model.ShoplistArticle
 import com.arribas.myshoppinglist.data.repository.ShoplistArticleRepository
 import com.arribas.myshoppinglist.data.repository.ShoplistRepository
 import com.arribas.myshoppinglist.data.utils.DIALOG_UI_TAG
 import com.arribas.myshoppinglist.data.utils.DialogUiState
-import com.arribas.myshoppinglist.data.utils.PreferencesEnum
-import com.arribas.myshoppinglist.data.utils.PreferencesManager
 import com.arribas.myshoppinglist.ui.navigation.route.Routes
 import com.arribas.myshoppinglist.ui.view.filter.GeneralFilterUiState
 import com.arribas.myshoppinglist.ui.view.screen.shoplist.ShoplistUiState
-import com.arribas.myshoppinglist.ui.view.screen.shoplist.shoplistList.ShoplisListUiState
+import com.arribas.myshoppinglist.ui.view.screen.shoplist.shoplistManage.ShoplistManageUiState
 import com.arribas.myshoppinglist.ui.view.screen.shoplist.toShopListUiState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
@@ -44,25 +41,20 @@ class ShoplistDetailViewModel(
     private val _dialogState = MutableStateFlow(DialogUiState())
     val dialogState: StateFlow<DialogUiState> = _dialogState
 
-    private val _listUiState = MutableStateFlow(ShoplistDetailUiState())
-    var listUiState: StateFlow<ShoplistDetailUiState> = _listUiState
+    private val _listUiState = MutableStateFlow(ShoplistManageUiState(context))
+    var listUiState: StateFlow<ShoplistManageUiState> = _listUiState
 
     private val _filterUiState = MutableStateFlow(GeneralFilterUiState())
     var filterUiState: StateFlow<GeneralFilterUiState> = _filterUiState
 
-    lateinit var article: ArticleShop
-
-    private val _shoplistListUiState = MutableStateFlow(ShoplisListUiState())
-    val shoplistListUiState: StateFlow<ShoplisListUiState> = _shoplistListUiState
-
-    lateinit var mode: ShoplistDetailModeEnum
+    lateinit var article: ShoplistArticle
 
     init{
         getData()
     }
 
     /**Public functions**************************************/
-    fun onUpdateItem(_article: ArticleShop) {
+    fun onUpdateItem(_article: ShoplistArticle) {
         article = _article
 
         if(article.quantity == 0){
@@ -76,123 +68,64 @@ class ShoplistDetailViewModel(
     fun onReorderItems(startIndex: Int, endIndex: Int){
         val articleTo = listUiState.value.itemList.get(startIndex)
         val articleFrom = listUiState.value.itemList.get(endIndex)
-    /*
+
         viewModelScope.launch(Dispatchers.IO) {
-            articleShopRepository.updateItem(articleTo.copy(order = articleFrom.order))
-            articleShopRepository.updateItem(articleFrom.copy(order = articleTo.order))
+            shoplistArticleRepository.updateItem(articleTo.copy(order = articleFrom.order))
+            shoplistArticleRepository.updateItem(articleFrom.copy(order = articleTo.order))
         }
-    */
-    }
-
-    fun onSearch(_name: String){
-        _filterUiState.value = _filterUiState.value.copy(name = _name)
-
-        getData()
-    }
-
-    fun onClearName(){
-        _filterUiState.value = _filterUiState.value.copy(name = "")
-        getData()
-    }
-
-    fun onDialogDelete(_article: ArticleShop){
-        article = _article
-
-        //onOpenDialogClicked(tag = DIALOG_UI_TAG.TAG_DELETE)
     }
 
     fun onDialogReset(){
-        //onOpenDialogClicked(tag = DIALOG_UI_TAG.TAG_RESET)
+        onOpenDialogClicked(tag = DIALOG_UI_TAG.TAG_RESET)
     }
 
-    fun setShoplist(element: ShoplistUiState){
-        shoplistUiState = shoplistUiState.copy(
-            id = element.id,
-            name = element.name,
-            type = element.type
-        )
+    fun onDialogDelete(_article: ShoplistArticle){
+        article = _article
 
-        PreferencesManager(context).saveData(
-            PreferencesEnum.MAIN_LIST.toString(),
-            this.shoplistUiState.id.toString()
-        )
+        onOpenDialogClicked(tag = DIALOG_UI_TAG.TAG_DELETE)
     }
 
     /**Private functions********************************************/
     private fun getData() {
-        viewModelScope.launch{
-            try {
-                viewModelScope.launch(Dispatchers.IO) {
-                    delay(100)
-                    shoplistUiState = shoplistRepository.getItem(itemId!!.toInt())
-                        .filterNotNull()
-                        .first()
-                        .toShopListUiState()
+        try {
+            viewModelScope.launch(Dispatchers.IO) {
+                delay(100)
+                shoplistUiState = shoplistRepository.getItem(itemId.toInt())
+                    .filterNotNull()
+                    .first()
+                    .toShopListUiState()
 
-                    shoplistArticleRepository.getItemsByList(itemId!!.toInt())
-                        .collect { list ->
-                            _listUiState.value = ShoplistDetailUiState(
-                                itemList = list,
-                                itemCount = list.count(),
-                                itemSelectCount = list.count { it.check }
-                            )
-                        }
-
-                }
-
-            }catch (e: IllegalArgumentException){
-
+                getDataShoplist()
             }
-            /*articleShopRepository.getAllItems()
-                .collect { list ->
-                    _listUiState.value = ListArticleShopUiState(
-                        itemList = list,
-                        itemCount = list.count(),
-                        itemSelectCount = list.count { it.check }
-                    )
-                }*/
+
+        }catch (e: IllegalArgumentException){
+
         }
     }
 
-    private fun getDataShoplist(){
-        viewModelScope.launch {
-            shoplistRepository.getItemsByName(filterUiState.value.name).collect {
-                    list ->
-                _shoplistListUiState.value = ShoplisListUiState(
+    private suspend fun getDataShoplist(){
+        shoplistArticleRepository.getItemsByList(itemId.toInt(), _filterUiState.value.name)
+            .collect { list ->
+                _listUiState.value = ShoplistManageUiState(
+                    context = context,
                     itemList = list,
-                    itemCount = list.count()
+                    itemCount = list.count(),
+                    itemSelectCount = list.count { it.check }
                 )
             }
-        }
-    }
-
-    private fun reset(){
-        viewModelScope.launch(Dispatchers.IO) {
-           // articleShopRepository.reset()
-        }
     }
 
     private fun updateItem(){
         viewModelScope.launch {
-          //  articleShopRepository.updateItem(article)
+            shoplistArticleRepository.updateItem(article)
         }
     }
 
     private fun deleteItem() {
         viewModelScope.launch {
-         /*   articleShopRepository.deleteItem(article)
+            shoplistArticleRepository.deleteItem(article)
 
-            val article = articleRepository.getItemByName(article.name);
-
-            article.first()?.let{ _article->
-                val article = _article.copy(
-                    shopCheked = false
-                )
-
-                articleRepository.updateItem(article)
-            }
-
-            updateOrderItems()*/
+            updateOrderItems()
         }
     }
 
@@ -201,23 +134,23 @@ class ShoplistDetailViewModel(
             var count = 1;
 
             listUiState.value.itemList.forEach { article ->
-               // articleShopRepository.updateItem(article.copy(order = count++))
+                shoplistArticleRepository.updateItem(article.copy(order = count++))
             }
         }
     }
 
     /**AlertDialog functions****************************************/
     fun onDialogConfirm() {
-       /* when(_dialogState.value.tag) {
-            DIALOG_UI_TAG.TAG_DELETE  -> deleteItem()
-            DIALOG_UI_TAG.TAG_RESET   -> reset()
-            DIALOG_UI_TAG.TAG_SUCCESS -> onDialogDismiss()
+       when(_dialogState.value.tag) {
+           DIALOG_UI_TAG.TAG_DELETE  -> deleteItem()
+           DIALOG_UI_TAG.TAG_RESET   -> reset()
+           DIALOG_UI_TAG.TAG_SUCCESS -> onDialogDismiss()
 
             else -> {}
-        }
+       }
 
         _dialogState.value = DialogUiState(
-            isShow = false)*/
+            isShow = false)
     }
 
     fun onDialogDismiss() {
@@ -227,10 +160,18 @@ class ShoplistDetailViewModel(
                     article.copy(quantity = 1)
                 }
             }
+            DIALOG_UI_TAG.TAG_RESET   -> reset()
+            DIALOG_UI_TAG.TAG_SUCCESS -> onDialogDismiss()
             else -> {}
         }
 
         _dialogState.value = DialogUiState(isShow = false)
+    }
+
+    private fun reset(){
+        viewModelScope.launch(Dispatchers.IO) {
+            shoplistArticleRepository.reset(shoplistUiState.id)
+        }
     }
 
     private fun onOpenDialogClicked(tag: DIALOG_UI_TAG) {
@@ -255,6 +196,7 @@ class ShoplistDetailViewModel(
                     isShowBtDismiss = true
                 )
 
+
             DIALOG_UI_TAG.TAG_SUCCESS ->
                 _dialog = DialogUiState(
                     tag = tag,
@@ -275,11 +217,7 @@ data class ShoplistDetailUiState(
     var itemList: List<ShoplistArticle> = listOf(),
     val itemCount: Int = 0,
     val itemSelectCount: Int = 0
-){
-    fun getTotalItemText(): String {
-        return "Total Items ${itemCount} - Select Items ${itemSelectCount}"
-    }
-}/*
+)/*
 data class ListArticleShopUiState(
     var itemList: List<ArticleShop> = listOf(),
     val itemCount: Int = 0,
